Did the restructuring
