+++ CMakeLists.txt
@@ -14,11 +14,17 @@ set(headers
         include/Reshade/reshade_api_resource.hpp
         include/Reshade/reshade_events.hpp
         include/Reshade/reshade_overlay.hpp
-        include/ReShadeToggler.h)
+        include/ReShadeToggler.h
+        include/Processor.h
+        include/ReshadeIntegration.h
+        include/Globals.h
+)
 
 
 set(sources 
             src/ReShadeToggler.cpp
+            src/Processor.cpp
+            src/ReshadeIntegration.cpp
 )
 
 # #
